Extract bracket view lifecycle module
Move bracket rendering and control state behind a one-time init and
repeatable activate/deactivate contract for the unified viewer.

Refs #270
